\name{eventChar.df}
\alias{eventChar.df}
\docType{data}
\title{
Event name - character conversion key
}
\description{
A data frme where each element in column event (event name) corresponds to an element in column char (character),
which can be a letter, digit, or a special character.
}
\usage{data("eventChar.df")}
\format{
  A data frame with 16 observations on the following 2 variables.
  \describe{
    \item{\code{event}}{a character vector}
    \item{\code{char}}{a character vector}
  }
}

\examples{
data(eventChar.df)

}
\keyword{datasets}
