\name{p1_20up}
\alias{p1_20up}
\docType{data}
\title{
Patterns from string group 1
}
\description{
Patterns that occur at least 20 percent compared to the number of strings 
in string group 1. It can be obtained from one of the exported files from
CommonPattern(str1).
}
\usage{data(p1_20up)}
\format{
  The format is:
 chr [1:32] "212" "202" "BAB" "D0D" "F0F" "020" "B0B" "010" "404" "C0C" ...
}

\examples{
data(p1_20up)

}
\keyword{datasets}
