\name{p2_25up}
\alias{p2_25up}
\docType{data}
\title{
Patterns from string group 2
}
\description{
Patterns that occur at least 25 percent compared to the number of strings 
in string group 2. It can be obtained from one of the exported files from
CommonPattern(str2).
}
\usage{data(p2_25up)}
\format{
  The format is:
 chr [1:32] "0D0D" "0E0E" "E0E0" "D0D" "E0E" "F0F" "B0B" "0C0" "0D0" ...
}

\examples{
data(p2_25up)

}
\keyword{datasets}
