\name{str1}
\alias{str1}
\docType{data}
\title{
String group 1
}
\description{
A vector containing 45 strings that have different lengths. It also can be 
obtained in the export file from the example in function EveString.
}
\usage{data("str1")}
\format{
  The format is:
 chr [1:45] "D02F0E20DEDC0C30BDC0E45G050A0B5050A06BG0BA5607BA" ...
}

\examples{
data(str1)

}
\keyword{datasets}
