\name{str2}
\alias{str2}
\docType{data}
\title{
String group 2
}
\description{
A vector containing 29 strings that have different lengths.
}
\usage{data(str2)}
\format{
  The format is:
 chr [1:29] "G21A1C14C2D0D21D2123201D23D21234320431212412421AB3EGEGE0E4G4B5G6A" ...
}

\examples{
data(str2)

}
\keyword{datasets}
