\name{event1s.df}
\alias{event1s.df}
\docType{data}
\title{
Data frame containing event names
}
\description{
A data frame containing event names, which are from an eye tracking study.
Thus, each event name is actually an area of interst (AOI). There are 45 rows.
Each row has 26 event names.
}
\usage{data("event1s.df")}
\format{
  A data frame with 45 observations or rows.
}

\examples{
data(event1s.df)

}
\keyword{datasets}
